# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/user/shop/:shop_id/etalase": {
            "get": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Get etalase shop user by its id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get etalase of a shop user.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The shop id",
                        "name": "shop_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Commerce_Etalase"
                            }
                        }
                    }
                }
            }
        },
        "/api/user/shop/:shop_id/etalase/:etalase_id": {
            "get": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Get a etalase shop by shop id and etalase id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get one etalase from one shop.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The shop id",
                        "name": "shop_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The etalase id",
                        "name": "etalase_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Commerce_Etalase"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Delete a shop user etalase by its id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Delete one user etalase.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The etalase id",
                        "name": "etalase_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The shop id",
                        "name": "shop_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Update a user shop etalase by shop id and etalase id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Update an user shop etalase.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The etalase id",
                        "name": "etalase_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The shop id",
                        "name": "shop_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "the body to update an etalase",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.UpdateEtalaseInput"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Commerce_Etalase"
                        }
                    }
                }
            }
        },
        "/api/user/shop/:shop_id/etalase/create": {
            "post": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Creating an etalase shop user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Create an etalase shop user.",
                "parameters": [
                    {
                        "description": "the body to create an etalase",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.CreateEtalaseInput"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Commerce_Etalase"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.CreateEtalaseInput": {
            "type": "object",
            "properties": {
                "etalase_name": {
                    "type": "string"
                },
                "shop_id": {
                    "type": "integer"
                }
            }
        },
...
        "controllers.UpdateEtalaseInput": {
            "type": "object",
            "properties": {
                "etalase_name": {
                    "type": "string"
                },
                "shop_id": {
                    "type": "integer"
                }
            }
        },
...
        "models.Commerce_Etalase": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "etalase_id": {
                    "type": "integer"
                },
                "etalase_name": {
                    "type": "string"
                },
                "shop_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    }
}